requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
razorpay==1.4.2
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
gunicorn==21.2.0
wheel==0.45.1
email-validator==2.1.1
uvicorn[standard]==0.34.3
jinja2==3.1.6





